feat: menuClickCallback api
